+++ examplepatches/PrathxmOp/Prathxm-Patches/extensions/extension/src/main/java/app/prathxm/chess/extension/lichesspuzzle/LichessPuzzleJourneyActivity.java
@@ -65,7 +65,7 @@ public class LichessPuzzleJourneyActivity extends Activity implements PuzzleJour
         super.onCreate(savedInstanceState);
         setupJourneyUI();
 
-        dbHelper = new LichessPuzzleDatabaseHelper(this);
+        dbHelper = LichessPuzzleDatabaseHelper.getInstance(this);
         downloadManager = new PuzzleDownloadManager(
                 this,
                 mainContainer,