+++ examplepatches/PrathxmOp/Prathxm-Patches/extensions/extension/src/main/java/app/prathxm/chess/extension/stockfish/LocalAnalysisFlow.java
@@ -16,10 +16,21 @@ import java.util.Map;
 public class LocalAnalysisFlow {
     private static final String TAG = "LocalAnalysisFlow";
 
+    private static Class<?> findClass(String... names) throws ClassNotFoundException {
+        for (String name : names) {
+            try {
+                return Class.forName(name);
+            } catch (ClassNotFoundException e) {
+                // Try next
+            }
+        }
+        throw new ClassNotFoundException("Could not find any of: " + java.util.Arrays.toString(names));
+    }
+
     public static Object createFlow(final String pgn, final Object analysisDepthObj) {
         try {
             ClassLoader classLoader = StockfishExtension.class.getClassLoader();
-            Class<?> g74Class = Class.forName("com.google.android.g74");
+            Class<?> g74Class = findClass("com.google.android.hb4", "com.google.android.g74");
 
             return Proxy.newProxyInstance(
                 classLoader,
@@ -73,8 +84,8 @@ public class LocalAnalysisFlow {
             Class<?> adClass = Class.forName("com.chess.entities.AnalysisDepth");
             Class<?> mClass = Class.forName("com.chess.gamereview.repository.m");
             Class<?> maClass = Class.forName("com.chess.gamereview.repository.m$a");
-            Class<?> a84Class = Class.forName("com.google.android.a84");
-            Class<?> o02Class = Class.forName("com.google.android.o02");
+            Class<?> a84Class = findClass("com.google.android.bc4", "com.google.android.a84");
+            Class<?> o02Class = findClass("com.google.android.i02", "com.google.android.o02");
 
             Method emitMethod = a84Class.getMethod("emit", Object.class, o02Class);
 
@@ -510,8 +521,8 @@ public class LocalAnalysisFlow {
             Log.e(TAG, "Local stockfish analysis failed", t);
             try {
                 Class<?> failureClass = Class.forName("com.chess.gamereview.repository.h$a");
-                Class<?> a84Class = Class.forName("com.google.android.a84");
-                Class<?> o02Class = Class.forName("com.google.android.o02");
+                Class<?> a84Class = findClass("com.google.android.bc4", "com.google.android.a84");
+                Class<?> o02Class = findClass("com.google.android.i02", "com.google.android.o02");
                 Method emitMethod = a84Class.getMethod("emit", Object.class, o02Class);
                 Constructor<?> failConstructor = failureClass.getConstructor(Throwable.class);
                 Object failureResult = failConstructor.newInstance(t);