+++ examplepatches/PrathxmOp/Prathxm-Patches/patches/src/main/kotlin/app/prathxm/chess/patches/shared/Constants.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright 2026 PrathxmOp
+ * https://github.com/PrathxmOp/Prathxm-Patches
+ */
+
 package app.prathxm.chess.patches.shared
 
 import app.morphe.patcher.patch.ApkFileType
@@ -8,11 +13,14 @@ object Constants {
     val COMPATIBILITY_CHESS = Compatibility(
         name = "Chess.com",
         packageName = "com.chess",
-        apkFileType = ApkFileType.APK,
+        apkFileType = ApkFileType.APKM,
         appIconColor = 0x81B64C, // Chess.com green
+        signatures = setOf("aa8c9ff93efbdc4226a113e49fc9645b2c02092a314305da2d76be6cdf8abfc3"),
         targets = listOf(
             AppTarget(version = "4.9.49"),
-            AppTarget(version = "4.9.49-googleplay")
+            AppTarget(version = "4.9.49-googleplay"),
+            AppTarget(version = "4.10.0"),
+            AppTarget(version = "4.10.0-googleplay")
         )
     )
 }